fim de todas as funcionalidades, só arrumar o programa

+++ src/main/java/controller/ControllerTelaViewLiga.java
@@ -144,6 +144,20 @@ public class ControllerTelaViewLiga {
             mostrarAlerta("Erro", "Falha ao carregar usuários da liga.");
             e.printStackTrace();
         }
+
+        listViewUsuarios.setOnMouseClicked(event -> {
+            if (event.getClickCount() == 2) {
+                if (!Simulacao.getOcorreu()) {
+                    mostrarAlerta("Simulação não realizada", "A simulação ainda não foi realizada. Você não pode visualizar os times.");
+                    return;
+                }
+
+                Usuario selecionado = listViewUsuarios.getSelectionModel().getSelectedItem();
+                if (selecionado != null) {
+                    abrirTelaVisualizarCampinho(selecionado);
+                }
+            }
+        });
     }
 
     @FXML
@@ -153,6 +167,14 @@ public class ControllerTelaViewLiga {
 
     @FXML
     private void sairLiga() {
+        if(Simulacao.getOcorreu()) {
+            Alert alerta = new Alert(Alert.AlertType.WARNING);
+            alerta.setTitle("Ação bloqueada");
+            alerta.setHeaderText(null);
+            alerta.setContentText("Você não pode sair da liga pois a simulação já foi realizada.");
+            alerta.showAndWait();
+            return;
+        }
         try {
 
             boolean sucesso = usuarioDAO.removerUsuarioDaLiga(usuario, usuario.getLiga());
@@ -186,6 +208,29 @@ public class ControllerTelaViewLiga {
             e.printStackTrace();
         } catch (SQLException e) {
             mostrarAlerta("Erro", "Erro no banco de dados ao sair da liga.");
+            e.printStackTrace();
+        }
+    }
+
+    private void abrirTelaVisualizarCampinho(Usuario outroUsuario) {
+        try {
+            FXMLLoader loader = new FXMLLoader(getClass().getResource("/screens/UsrLigaScreens/TelaVisualizarCampinho.fxml"));
+            Parent root = loader.load();
+
+            ControllerTelaVisualizarCampinho controller = loader.getController();
+            controller.setConnection(conn, outroUsuario);
+
+            Stage stage = (Stage) menuMontagem.getScene().getWindow();
+
+            SceneInfo sceneInfo = new SceneInfo(menuMontagem.getScene(), stage.getTitle());
+            NavigationManager.push(sceneInfo);
+
+            stage.setScene(new Scene(root));
+            stage.setTitle("Visualizar Time de " + outroUsuario.getNome());
+            stage.show();
+
+        } catch (IOException e) {
+            mostrarAlerta("Erro", "Erro ao abrir visualização do time.");
             e.printStackTrace();
         }
     }
